+++ vscode/src/chat/MessageProvider.ts
@@ -105,10 +105,6 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
     constructor(options: MessageProviderOptions) {
         super()
 
-        if (TestSupport.instance) {
-            TestSupport.instance.messageProvider.set(this)
-        }
-
         this.chat = options.chat
         this.intentDetector = options.intentDetector
         this.guardrails = options.guardrails
@@ -367,12 +363,7 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
 
         // Filter the human input to check for chat commands and retrieve the correct recipe id
         // e.g. /edit from 'chat-question' should be redirected to use the 'fixup' recipe
-        const command = await this.chatCommandsFilter(
-            humanChatInput,
-            recipeId,
-            { source, requestID },
-            userInputContextFiles
-        )
+        const command = await this.chatCommandsFilter(humanChatInput, recipeId, { source, requestID })
         if (!command) {
             return
         }
@@ -576,14 +567,13 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
                 break
         }
         // Get prompt details from controller by title then execute prompt's command
-        return this.executeRecipe('custom-prompt', title, 'custom-commands')
+        return vscode.commands.executeCommand('cody.action.commands.exec', title)
     }
 
     protected async chatCommandsFilter(
         text: string,
         recipeId: RecipeID,
-        eventTrace?: { requestID?: string; source?: ChatEventSource },
-        userContextFiles?: ContextFile[]
+        eventTrace?: { requestID?: string; source?: ChatEventSource }
     ): Promise<{ text: string; recipeId: RecipeID; source?: ChatEventSource } | void> {
         const source = eventTrace?.source || undefined
         text = text.trim()
@@ -634,23 +624,7 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
                     const assistantResponse = 'Command failed. Please open a file and try again.'
                     return this.addCustomInteraction({ assistantResponse, text, source })
                 }
-                const commandRunnerID = await this.editor.controllers.command?.addCommand(
-                    text,
-                    eventTrace?.requestID,
-                    userContextFiles
-                )
-                // no op
-                if (!commandRunnerID) {
-                    return
-                }
-
-                if (commandRunnerID === 'invalid') {
-                    const assistantResponse = `__${text}__ is not a valid command`
-                    // If no command found, send error message to view
-                    return this.addCustomInteraction({ assistantResponse, text, source })
-                }
-
-                return { text: commandRunnerID, recipeId: 'custom-prompt', source }
+                return vscode.commands.executeCommand('cody.action.commands.exec', text)
             }
         }
     }